% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_fetch_data.R
\docType{data}
\name{alt_colname_map}
\alias{alt_colname_map}
\title{Map for selecting and renaming football-data.co.uk columns}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 2 columns.
}
\usage{
alt_colname_map
}
\description{
Map for selecting and renaming football-data.co.uk columns
}
\keyword{internal}
