% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_lookup.R
\docType{data}
\name{alt_country_lookup}
\alias{alt_country_lookup}
\title{Lookup alternative football-data.co.uk country codes by name}
\format{
An object of class \code{list} of length 16.
}
\usage{
alt_country_lookup
}
\description{
Lookup alternative football-data.co.uk country codes by name
}
\keyword{internal}
