% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_fetch_data.R
\name{alt_fetch_data}
\alias{alt_fetch_data}
\title{Fetch data for a single country and season}
\usage{
alt_fetch_data(country, season)
}
\arguments{
\item{country}{The country of the league of interest}

\item{season}{The start-year of the season}
}
\description{
Fetch data for a single country and season
}
\examples{

# Fetching Danish data for 2014/15 season
alt_fetch_data("Denmark", 2014)

}
