% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_fetch_data.R
\name{alt_football_data_url}
\alias{alt_football_data_url}
\title{Get the url for the alternative leagues}
\usage{
alt_football_data_url(country)
}
\description{
Get the url for the alternative leagues
}
\keyword{internal}
