% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_fetch_data.R
\name{alt_season_code}
\alias{alt_season_code}
\title{Create a filter string for a given season}
\usage{
alt_season_code(start_year)
}
\description{
Create a filter string for a given season
}
\keyword{internal}
