% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\docType{data}
\name{col_spec}
\alias{col_spec}
\title{Column specification for football-data.co.uk csv files}
\format{
An object of class \code{col_spec} of length 2.
}
\usage{
col_spec
}
\description{
Column specification for football-data.co.uk csv files
}
\keyword{internal}
