% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\docType{data}
\name{colname_map}
\alias{colname_map}
\title{Map for selecting and renaming foorball-data.co.uk columns}
\format{An object of class \code{list} of length 25.}
\usage{
colname_map
}
\description{
Map for selecting and renaming foorball-data.co.uk columns
}
\keyword{internal}
