% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_lookup.R
\docType{data}
\name{country_lookup}
\alias{country_lookup}
\title{Lookup football-data.co.uk country codes by name}
\format{An object of class \code{list} of length 10.}
\usage{
country_lookup
}
\description{
Lookup football-data.co.uk country codes by name
}
\keyword{internal}
