% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data}
\usage{
fetch_data(country, division, season)
}
\arguments{
\item{country}{The country of the league of interest}

\item{division}{An integer denoting the division of interest, where lower numbers refer
to higher divisions. English and Scottish Premierships are `0`, while all other
top divisions (e.g. La Liga) are `1`.}

\item{season}{The start-year of the season}
}
\description{
Fetch data for a single league and season
}
\examples{

# Fetching Premier League data for 2014/15 season
fetch_data("England", 0, 2014)

# Fetching Championship data
fetch_data("England", 1, 2014)

# Fetching Serie A data
fetch_data("Italy", 1, 2014)

}
