% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{football_data_url}
\alias{football_data_url}
\title{Get the url for a given league and season}
\usage{
football_data_url(country, division, season)
}
\description{
Get the url for a given league and season
}
\keyword{internal}
