% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{season_code}
\alias{season_code}
\title{Get the football-data.co.uk string for a given season}
\usage{
season_code(start_year)
}
\description{
Get the football-data.co.uk string for a given season
}
\keyword{internal}
